requirements.txt
Django~=2.1
djangorestframework==3.8.2
djangorestframework-jwt--1.11.0
coreapi--2.3.3
psycopg2--2.7.5